Update JWT Security and Add DB Seeder

+++ Back-End/src/main/java/com/future/tcfm/service/impl/UserServiceImpl.java
@@ -1,4 +1,3 @@
-/*
 package com.future.tcfm.service.impl;
 
 import com.future.tcfm.model.User;
@@ -35,7 +34,7 @@ public class UserServiceImpl implements UserService {
             return new ResponseEntity<>("Failed to save User!\nGroup can't be null!", HttpStatus.BAD_REQUEST);
 
         user.setPassword(passwordEncoder.encode(user.getPassword()));//ENCRYPTION PASSWORD
-        user.setBalance(0);//FOR HANDLING NOT NULL PARAMATER
+        user.setBalance((double) 0);//FOR HANDLING NOT NULL PARAMATER
         userRepository.save(user);
         return new ResponseEntity<>(user, HttpStatus.OK);
     }
@@ -59,7 +58,6 @@ public class UserServiceImpl implements UserService {
         return new ResponseEntity<>(userExist, HttpStatus.OK);
     }
 }
-*/
 
 /*    @Override
     public ResponseEntity<?> deleteUser(String id) {
